\name{qkiosk-package}
\alias{qkiosk-package}
\alias{qkiosk}
\alias{qk_fund}
\alias{qk_univ}
\docType{package}
\title{
\packageTitle{qkiosk}
}
\description{
\packageDescription{qkiosk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qkiosk}
\packageIndices{qkiosk}

Official QUANTkiosk R API
}
\author{
\packageAuthor{qkiosk}

Maintainer: \packageMaintainer{qkiosk}
}
\references{
\url{http://quantkiosk.com}
}
\keyword{ package }
